Close #322 - CanRecover trait from effectie-cats-effect, effectie-cats-effect3 and effectie-monix
