+++ cats-effect/src/main/scala-2/effectie/cats/CanRecover.scala
@@ -1,29 +1,16 @@
 package effectie.cats
 
 import cats._
-import cats.data.EitherT
 import cats.effect._
 
-import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
 
 /** @author Kevin Lee
   * @since 2020-08-17
   */
-trait CanRecover[F[_]] extends effectie.CanRecover[F] {
-
-  type XorT[A, B] = EitherT[F, A, B]
-
-  @inline override final protected def xorT[A, B](fab: F[Either[A, B]]): EitherT[F, A, B] =
-    EitherT(fab)
-
-  @inline override final protected def xorT2FEither[A, B](efab: EitherT[F, A, B]): F[Either[A, B]] =
-    efab.value
-}
-
 object CanRecover {
 
-  def apply[F[_]: CanRecover]: CanRecover[F] = implicitly[CanRecover[F]]
+  type CanRecover[F[_]] = effectie.CanRecover[F]
 
   implicit object IoCanRecover extends CanRecover[IO] {
 
@@ -36,10 +23,6 @@ object CanRecover {
       recoverFromNonFatalWith[A, AA](fa)(handleError.andThen(IO.pure(_)))
 
   }
-
-  @SuppressWarnings(Array("org.wartremover.warts.ImplicitParameter"))
-  implicit def futureCanRecover(implicit ec: ExecutionContext): CanRecover[Future] =
-    new effectie.CanRecover.FutureCanRecover(ec) with CanRecover[Future]
 
   implicit object IdCanRecover extends CanRecover[Id] {
