+++ cats-effect/src/main/scala-2/effectie/cats/package.scala
@@ -32,6 +32,24 @@ package object cats {
 
   }
 
+  implicit final class CanRecoverOps[F[_]](private val canRecover: effectie.CanRecover[F]) extends AnyVal {
+
+    final def recoverEitherTFromNonFatalWith[A, AA >: A, B, BB >: B](
+      efab: => EitherT[F, A, B]
+    )(
+      handleError: PartialFunction[Throwable, F[Either[AA, BB]]]
+    ): EitherT[F, AA, BB] =
+      EitherT(canRecover.recoverFromNonFatalWith[Either[A, B], Either[AA, BB]](efab.value)(handleError))
+
+    final def recoverEitherTFromNonFatal[A, AA >: A, B, BB >: B](
+      efab: => EitherT[F, A, B]
+    )(
+      handleError: PartialFunction[Throwable, Either[AA, BB]]
+    ): EitherT[F, AA, BB] =
+      EitherT(canRecover.recoverFromNonFatal[Either[A, B], Either[AA, BB]](efab.value)(handleError))
+
+  }
+
   implicit final class FxhOps[F[_]](private val fx: Fx[F]) extends AnyVal {
 
     def catchNonFatalEitherT[A, AA >: A, B](fab: => EitherT[F, A, B])(f: Throwable => AA): EitherT[F, AA, B] =
