+++ cats-effect/src/main/scala-3/effectie/cats/syntax/error.scala
@@ -1,7 +1,7 @@
 package effectie.cats.syntax
 
 import cats.data.EitherT
-import effectie.{CanCatch, CanHandleError}
+import effectie.{CanCatch, CanHandleError, CanRecover}
 import effectie.cats.*
 
 /** @author Kevin Lee
@@ -116,14 +116,14 @@ trait error {
     )(
       using canRecover: CanRecover[F]
     ): EitherT[F, AA, BB] =
-      canRecover.recoverEitherTFromNonFatalWith[A, AA, B, BB](efab)(handleError)
+      effectie.cats.recoverEitherTFromNonFatalWith(canRecover)[A, AA, B, BB](efab)(handleError)
 
     def recoverEitherTFromNonFatal[AA >: A, BB >: B](
       handleError: PartialFunction[Throwable, Either[AA, BB]]
     )(
       using canRecover: CanRecover[F]
     ): EitherT[F, AA, BB] =
-      canRecover.recoverEitherTFromNonFatal[A, AA, B, BB](efab)(handleError)
+      effectie.cats.recoverEitherTFromNonFatal(canRecover)[A, AA, B, BB](efab)(handleError)
   }
 
 }
