+++ cats-effect3/src/main/scala-2/effectie/cats/syntax/error.scala
@@ -2,8 +2,8 @@ package effectie.cats.syntax
 
 import cats.data.EitherT
 import effectie.cats.syntax.error.{EitherTFABErrorHandlingOps, FAErrorHandlingOps, FEitherABErrorHandlingOps}
-import effectie.cats.{CanCatchOps, CanHandleErrorOps, CanRecover}
-import effectie.{CanCatch, CanHandleError}
+import effectie.cats.{CanCatchOps, CanHandleErrorOps, CanRecoverOps}
+import effectie.{CanCatch, CanHandleError, CanRecover}
 
 /** @author Kevin Lee
   * @since 2021-10-16
