+++ cats-effect3/src/test/scala-3/effectie/cats/syntax/errorSpec.scala
@@ -2744,6 +2744,7 @@ object CanRecoverSyntaxSpec {
 
   object IOSpec {
     import effectie.cats.Fx.given
+    import effectie.cats.CanRecover.given
 
     def testCanRecover_IO_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
 
@@ -3771,6 +3772,7 @@ object CanRecoverSyntaxSpec {
 
   object IdSpec {
     import effectie.cats.Fx.given
+    import effectie.cats.CanRecover.given
 
     def testCanRecover_Id_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
